ensure that index name is never empty

+++ couchbase/operations/management/search_index_drop.cxx
@@ -27,6 +27,9 @@ namespace couchbase::operations::management
 std::error_code
 search_index_drop_request::encode_to(encoded_request_type& encoded, http_context& /* context */) const
 {
+    if (index_name.empty()) {
+        return error::common_errc::invalid_argument;
+    }
     encoded.method = "DELETE";
     encoded.path = fmt::format("/api/index/{}", index_name);
     return {};
